bugs resolved

+++ app/Livewire/Admin/IncludeManager/Index.php
@@ -166,8 +166,11 @@ class Index extends Component
     {
         $delid = $data['inputAttributes']['delid'];
         $model = IncludeManager::find($delid);
-        $uploadImageId = $model->includeManagerImage->id;
-        deleteFile($uploadImageId);
+        $uploadImageId = null;
+        if ($model->includeManagerImage) {
+            $uploadImageId = $model->includeManagerImage->id;
+            deleteFile($uploadImageId);
+        }
         $model->delete();
         $this->flash('success',  getLocalization('delete_success'));
     }
